innovation/layoutlm_pdf/requirements.txt
# LayoutLMv3 stack — CPU-only viable

# Transformers (LayoutLMv3)
transformers>=4.40.0

# PyTorch CPU
# pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0

# Image handling
Pillow>=10.0.0

# PDF → images conversion (requires poppler: apt-get install poppler-utils)
pdf2image>=1.16.0

# OCR (requires tesseract: apt-get install tesseract-ocr tesseract-ocr-deu tesseract-ocr-fra tesseract-ocr-spa)
pytesseract>=0.3.10

# Optional: PyMuPDF for plain-text fallback when pdf2image/tesseract unavailable
# PyMuPDF>=1.23.0

# Testing
pytest>=7.0.0
pytest-mock>=3.12.0
